gfx1030 (RDNA2): avoid fdot2 lowering and 32-bit warp shuffle masks
Two RDNA2-only kernel fixes, both no-ops on RDNA3/4 and NVIDIA:

- causal_conv1d: an all-bf16 multiply-accumulate lets LLVM fold the
  reduction into llvm.amdgcn.fdot2.bf16.bf16, which does not exist on
  gfx1030 (RDNA2) - the compiler hard-crashes (LLVM ERROR: Cannot
  select) at kernel compile time. Compute the taps in fp32; the
  accumulator was already fp32, so results are unchanged on other
  targets.
- the borrowed ggml warp-shuffle wrappers passed a 32-bit lane mask,
  which HIP rejects on wave64 targets (and is UB for lanes 32-63).
  Cast the mask to 64-bit and pin the butterfly width to 32 lanes so
  the reduction semantics match the CUDA originals under either wave
  size.

+++ python/freetoken/kernel/triton/causal_conv1d_triton.py
@@ -174,7 +174,8 @@ def _causal_conv1d_fwd_tiled_kernel(
         )
         mask_x = (src_local >= 0)[:, None] & (src_local < seqlen)[:, None] & mfc
         xj = tl.load(x_ptrs, mask_x, 0.0)
-        acc += w_j[None, :] * xj
+        # fp32 mul-add: all-bf16 fma folds to fdot2.bf16 (unsupported on gfx1030)
+        acc += w_j[None, :].to(tl.float32) * xj.to(tl.float32)
         if HAS_INITIAL_STATES:
             if load_init_state:
                 st_idx = state_len + src_local        # 0..state_len-1 where src_local<0
@@ -184,7 +185,7 @@ def _causal_conv1d_fwd_tiled_kernel(
                 )
                 mask_s = (src_local < 0)[:, None] & mfc
                 sj = tl.load(s_ptrs, mask_s, 0.0)
-                acc += w_j[None, :] * sj
+                acc += w_j[None, :].to(tl.float32) * sj.to(tl.float32)
 
     if SILU_ACTIVATION:
         # silu(x)=x/(1+exp(-x)); exp2 lowers to the native ex2.approx SFU op (faster than the
@@ -235,7 +236,7 @@ def _causal_conv1d_fwd_tiled_kernel(
                         & ((idx_tok + seqlen) < state_len)[:, None]
                         & mfc
                     )
-                    old = tl.load(src_ptrs, mask_src, 0.0)
+                    old = tl.load(src_ptrs, mask_src, 0.0).to(new_cs.dtype)
                     new_cs = tl.where(mask_src, old, new_cs)
         tgt = conv_states_base[None, :] + (idx_tok * stride_conv_state_tok)[:, None]
         mask_t = (idx_tok < state_len)[:, None] & mfc
@@ -392,7 +393,9 @@ def _causal_conv1d_update_kernel(
                 elif j == 3:
                     matrix_w = w_col3
                     matrix_x = tl.load(x_base_1d + idx_token * stride_x_token, mask=mask_x_1d)
-            acc += matrix_x * matrix_w
+            # fp32 mul-add: an all-bf16 fma lets LLVM fold it into
+            # llvm.amdgcn.fdot2.bf16, which does not exist on gfx1030 (RDNA2)
+            acc += matrix_x.to(tl.float32) * matrix_w.to(tl.float32)
 
         if KERNEL_WIDTH == 2:
             col0 = matrix_x
